LPS-64471 split upgrade process in 2 in order to meassure better if there is a performance issue
